falta para cambiar tileset

+++ Kings_Valley-parent/core/src/main/java/modelo/Mummy.java
@@ -9,7 +9,7 @@ public class Mummy extends GameCharacter
 {
 
     public Mummy(float x, float y, int p0, 
-	    int speedFall, int speedWalk, int speedWalkStairs, int speedJump,Pyramid pyramid)
+	    float speedFall, float speedWalk, float speedWalkStairs, float speedJump,Pyramid pyramid)
     {
 	super(Constantes.It_mummy, x, y, p0, 0, Config.getInstance().getCharacterWidth(), Config.getInstance().getCharacterHeight(), speedFall, speedWalk, speedWalkStairs, speedJump,pyramid);
